Hice una actualizacion para los estados CRUD

+++ src/main/java/com/connectEdu/crudProject/repository/RecursoRepsoitory.java
@@ -1,9 +1,10 @@
 package com.connectEdu.crudProject.repository;
 
-import org.springframework.data.jpa.repository.JpaRepository;
+
 
 import com.connectEdu.crudProject.model.Recurso;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RecursoRepsoitory extends JpaRepository<Recurso, Integer>{
-    
+public interface RecursoRepsoitory extends JpaRepository<Recurso, Long> {
+
 }
